requirements.txt
streamlit
PyGithub
openai
python-dotenv